// apps/nfid-frontend/src/integration/_ic_api/internet_identity.d.ts
export type AddTentativeDeviceResponse =
  | {
      device_registration_mode_off: null
    }
  | { another_device_tentatively_added: null }
  | {
      added_tentatively: {
        verification_code: string
        device_registration_timeout: Timestamp
      }
    }
export interface Challenge {
  png_base64: string
  challenge_key: ChallengeKey
}
export type ChallengeKey = string
export interface ChallengeResult {
  key: ChallengeKey
  chars: string
}
export type CredentialId = Array<number>
export interface Delegation {
  pubkey: PublicKey
  targets: [] | [Array<Principal>]
  expiration: Timestamp
}
export interface DeviceData {
  alias: string
  protection: DeviceProtection
  pubkey: DeviceKey
  key_type: KeyType
  purpose: Purpose
  credential_id: [] | [CredentialId]
}
export type DeviceKey = PublicKey
export type DeviceProtection = { unprotected: null } | { protected: null }
// ...
export type KeyType =
  | { platform: null }
  | { seed_phrase: null }
  | { cross_platform: null }
  | { unknown: null }
export type PublicKey = Array<number>
export type Purpose = { authentication: null } | { recovery: null }
// ...
export type Timestamp = bigint